% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myspotifymood.R
\name{my_spotify_mood}
\alias{my_spotify_mood}
\title{Visualizes Valence From Spotify}
\usage{
my_spotify_mood(username, client_id, client_secret, playlists)
}
\arguments{
\item{username}{The Spotify username of the individual utilizing this
function.}

\item{client_id}{The user's Spotify Developer Client ID}

\item{client_secret}{The user's Spotify Developer Client Secret.}

\item{playlists}{The user's chosen playlist's id.}
}
\value{
A visualization.
}
\description{
A function that allows the user to retrieve the individual tracks valance
(happiness level) of a specific playlist.
}
\examples{
myspotifymood("billybob08", "cbfj2895nbdudnejf2", "2kjgkjbrgrslenfjkrnjsne",
"34IdsklfjbejQkbrsFG458HSf")

}
\author{
Alexa Pope
}
